requirements.txt
numpy
jupyter
tensorflow-gpu
keras
pandas
numpy
pyyaml
psutil
matplotlib
tables
